Remove old test cases
