+++ zetaclient/orchestrator/v2_bootstrap_test.go
@@ -120,6 +120,46 @@ func TestBootstrap(t *testing.T) {
 	})
 }
 
+func TestBtcDatabaseFileName(t *testing.T) {
+	tests := []struct {
+		name     string
+		chain    chains.Chain
+		expected string
+	}{
+		{
+			name:     "should use legacy file name for bitcoin mainnet",
+			chain:    chains.BitcoinMainnet,
+			expected: "btc_chain_client",
+		},
+		{
+			name:     "should use legacy file name for bitcoin testnet3",
+			chain:    chains.BitcoinTestnet,
+			expected: "btc_chain_client",
+		},
+		{
+			name:     "should use new file name for bitcoin regtest",
+			chain:    chains.BitcoinRegtest,
+			expected: "btc_chain_client_btc_regtest",
+		},
+		{
+			name:     "should use new file name for bitcoin signet",
+			chain:    chains.BitcoinSignetTestnet,
+			expected: "btc_chain_client_btc_signet_testnet",
+		},
+		{
+			name:     "should use new file name for bitcoin testnet4",
+			chain:    chains.BitcoinTestnet4,
+			expected: "btc_chain_client_btc_testnet4",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			assert.Equal(t, tt.expected, btcDatabaseFileName(tt.chain))
+		})
+	}
+}
+
 func tasksHaveGroup(t *testing.T, tasks map[uuid.UUID]*scheduler.Task, group string) {
 	var found bool
 	for _, task := range tasks {
